chore: Initial Commit
The Terraform exporter for SAP BTP.

The SAP BTP Terraform provider has started gaining traction. In order for customers to use the terraform provider to manage their artifacts on BTP, they need to import these artifacts to terraform and write the configuration files for them.

This is a tedious task and prone to errors. Hence the btptfexporter will automatically generate the configuration and import the artifacts into the terraform state. As a result the customers will have a ready-made configuration to start wit as well as all the artifacts will be imported into terraform.

A new cli based tool that has basic set of flags to import to import some of the common artifacts from a Subaccount/Globalacount.